requirements.txt
streamlit
pypdf
google-generativeai
python-dotenv
pandas
